requirements.txt
# SmartSalesAgent Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WhatsApp Integration
twilio==8.10.0

# AI & NLP
openai==1.3.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Dashboard
streamlit==1.28.2
pandas>=2.0.0

# PDF Generation
reportlab==4.0.7

# Environment & Config
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
python-multipart==0.0.6

# Optional: For production deployment
# gunicorn==21.2.0
# redis==5.0.1
# celery==5.3.4
